pyproject.toml
[tool.poetry]
name = "python-airline-support-bot"
version = "0.1.0"
description = "An AI-powered airline support bot that can handle flight-related questions"
authors = ["Adam Nolte <ajnolte12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.76.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
click = "^8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
airline-bot = "python_airline_support_bot.cli:main"
